move checksum logic to checksum package
