+++ packages/checksum/index.js
@@ -1,4 +1,1 @@
-
-module.exports = {
-
-}
+module.exports = require('./checksum');
