+++ packages/checksum/package.json
@@ -32,8 +32,9 @@
   "author": "Cumulus Authors",
   "license": "Apache-2.0",
   "dependencies": {
-    "@cumulus/common": "1.11.2"
+    "cksum": "^1.3.0"
   },
   "devDependencies": {
+    "ava": "^0.25.0"
   }
 }
