+++ packages/checksum/tests/test-checksum.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const test = require('ava');
+
+test.serial('checksumFileStream returns checksum for file stream', () => {
+  return null;
+});
