+++ packages/common/index.js
@@ -6,7 +6,6 @@ exports.cliUtils = require('./cli-utils');
 exports.CloudFormationGateway = require('./CloudFormationGateway');
 exports.CollectionConfigStore = require('./collection-config-store').CollectionConfigStore;
 exports.constructCollectionId = require('./collection-config-store').constructCollectionId;
-exports.file = require('./file');
 exports.http = require('./http');
 exports.log = require('./log');
 exports.stepFunctions = require('./step-functions');
