+++ packages/common/package.json
@@ -41,12 +41,12 @@
   "author": "Cumulus Authors",
   "license": "Apache-2.0",
   "dependencies": {
+    "@cumulus/checksum": "1.11.2",
     "@cumulus/logger": "1.11.0",
     "@cumulus/test-data": "1.11.0",
     "ajv": "^5.2.2",
     "async": "^2.0.0",
     "aws-sdk": "^2.238.1",
-    "cksum": "^1.3.0",
     "expect.js": "^0.3.1",
     "follow-redirects": "^1.2.4",
     "formidable": "^1.2.1",
